// orders/src/module/orders/interfaces/orders.interface.ts
export const OrderStatus = {
  Created: 'Created',
  Confirmed: 'Confirmed',
  Processing: 'Processing',
  Collected: 'Collected',
  Done: 'Done',
  Canceled: 'Canceled',
  UNRECOGNIZED: 'UNRECOGNIZED',
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];
// ...
  export type Created = typeof OrderStatus.Created;
  export type Confirmed = typeof OrderStatus.Confirmed;
  export type Processing = typeof OrderStatus.Processing;
  export type Collected = typeof OrderStatus.Collected;
  export type Done = typeof OrderStatus.Done;
  export type Canceled = typeof OrderStatus.Canceled;
  export type UNRECOGNIZED = typeof OrderStatus.UNRECOGNIZED;
// ...
export interface Item {
  productId: string;
  quantity: number;
  unitPrice: number;
  sum: number;
}

export interface OrderResponse {
  orderId: string;
  status: OrderStatus;
  date: number;
  totalSum: number;
  customerId: string;
  items: Item[];
  createdAt: number;
}

export interface FindAllOrderResponse {
  orderId: string;
  status: OrderStatus;
  date: number;
  totalSum: number;
  customerId: string;
  createdAt: number;
}